requirements.txt
alembic==1.4.3
amqp==5.0.1
billiard==3.6.3.0
celery==5.0.1
certifi==2020.6.20
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cmake==3.18.4.post1
cycler==0.10.0
distro==1.5.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
importlib-metadata==2.0.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.2.0
kombu==5.0.2
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.19.2
opencv-python==4.4.0.46
packaging==20.4
pandas==1.1.4
Pillow==8.0.0
pkg-resources==0.0.0
prompt-toolkit==3.0.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
rq==1.6.1
scikit-build==0.11.1
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.20
threadpoolctl==2.1.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.3.1
